experimental addition of fwupd automatic
This only runs on AC. There is no battery state detection currently, battery should be at least at 70% (should cover most "battery saving" limits too).

Question is, if fwupdmgr works like that, or if it automatically reboots or something. If not, this is an essential step for reliable and automatic firmware updates. Maybe a preset "I have old firmware without updates" ?