refactor: Rewrite luks tpm2 script. (#302)
* refactor: Rewrite luks tpm2 script.

The previous version of the tpm2 script did not fail out when
encountering unsupported configurations. This script is narrowly focused
at enabling tpm2 decryption for root devices. By default we expect this
to be configured via the installer.

This script will now fail out if it cannot find the disk. The disk is
determined via the kernel commandline instead of /etc/crypttab. It still
only supports a single disk. It now also provides hints on why the
script has failed and the commands to run to disable and/or enable.

* refactor: Provide more information with script

* refactor: Include more safety checks

* fix: remove unnecessary luks check

---------

Co-authored-by: Benjamin Sherman <benjamin@holyarmy.org>