feat(just): Use `run0` instead of `sudo` in scripts
It is more secure than `sudo` & it uses native password dialog.

`run0` currently always asks for the password per root-call at the moment, so I modified the scripts which used `sudo` to initiate 1 call only. So user will only see 1 password prompt.
This may make some scripts slightly less readable, but it's still easy to read imo.

I added `librun0`, which only uses `run0` if it's installed, while fallbacking to `sudo` if not.
This makes sure that <=F40 images won't be affected with this change.

RPM is updated to reflect this change.